test: move test cases
