+++ packages/zilliqa-js-account/test/account.spec.ts
@@ -15,7 +15,6 @@
 //  You should have received a copy of the GNU General Public License
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import fs from 'fs';
 import Signature from 'elliptic/lib/elliptic/ec/signature';
 import { HTTPProvider } from '@zilliqa-js/core';
 import * as zcrypto from '@zilliqa-js/crypto';
@@ -48,17 +47,30 @@ describe('Account', () => {
   });
 
   it('should create accounts from keystores', async () => {
-    const content = await new Promise((resolve, reject) => {
-      fs.readFile(
-        `${__dirname}/fixtures/keystores_redux.json`,
-        'utf8',
-        (err, data) => {
-          err ? reject(err) : resolve(data);
-        },
-      );
-    });
+    const keystoreFixtures = [
+      {
+        privateKey:
+          '50af9e8c08011953f568d1818ad500a84b906117fb79502fc069525d1475d543',
+        passphrase: 'strong_password',
+        keystore:
+          '{"address":"0x3591c6b333776a5b6c885Ada53b4462dEe8c67B6","crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"1b23d4a09fa241b9cbe176fbc7d9567f"},"ciphertext":"40cce0a0b4eee4e4258341e18d880ff4b12ed25887252b90b7de4b6be86da5d5","kdf":"scrypt","kdfparams":{"salt":"4997f389a370424072967f6faeb709c238ff3cc3606a4101fc3dcbdd74b67108","n":8192,"c":262144,"r":8,"p":1,"dklen":32},"mac":"97f0677a9f0ad0c6052896e1b5ae94f86025047b9fe2617c1fce220af0db0e82"},"id":"39626239-3338-4138-b662-363663333030","version":3}',
+      },
+      {
+        privateKey:
+          '629fd0230abafb7bc8d575c604e6a1b3fabb2c995d5d6d0ab504156ce22d7136',
+        passphrase: 'strong_password',
+        keystore:
+          '{"address":"0x8d991C6222F26B9aB960dbD28A1e9f29ED77272d","crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"6177b7d2561056e9be3c991cce5148ce"},"ciphertext":"562c10f06d4fd050edcc4f99ef3390da34174c9d5a22ca46f4ed7d8ff0131e34","kdf":"scrypt","kdfparams":{"salt":"6c35d2b1918ff020bb52b4374105d9c5f9c20e586f2b4067818fe55ccc21372f","n":8192,"c":262144,"r":8,"p":1,"dklen":32},"mac":"b19578d520eac69a989e196fa65b059be8dd5644e157d065a17379f85d2a2994"},"id":"66373132-6463-4634-b936-356330326230","version":3}',
+      },
+      {
+        privateKey:
+          'c6525c60b970082edf830ea2ebdbe2437bdf715e0bf7e3fbb8d4b0e73ec029bc',
+        passphrase: 'strong_password',
+        keystore:
+          '{"address":"0xF5D26831Be0B1fe5a6Ba5286DC59FaacE63f24aD","crypto":{"cipher":"aes-128-ctr","cipherparams":{"iv":"44d6acb359150e9019e1dda58da2ceca"},"ciphertext":"3e37afff8624b000591dd84745738c64c2728b94a74da60d8600fa9419381694","kdf":"scrypt","kdfparams":{"salt":"bbf3723986c6ce90f8fe7846d2bc0bef40885688bef01733158652019f7f6c22","n":8192,"c":262144,"r":8,"p":1,"dklen":32},"mac":"6d8f5522a520dba0e6820715a4ae62dbc6412ae6d25aa3d2eacd8941f7c427a9"},"id":"35313363-3863-4265-b135-366531343061","version":3}',
+      },
+    ];
 
-    const keystoreFixtures = JSON.parse(content as string);
     for (const keystoreFixture of keystoreFixtures) {
       const { privateKey, passphrase, keystore } = keystoreFixture;
       const account = await Account.fromFile(keystore, passphrase);
