--- packages/zilliqa-js-account/test/fixtures/keystores_redux.json
@@ -1,52 +1,0 @@
-[
-  {
-    "privateKey": "50af9e8c08011953f568d1818ad500a84b906117fb79502fc069525d1475d543",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0x3591c6b333776a5b6c885Ada53b4462dEe8c67B6\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"1b23d4a09fa241b9cbe176fbc7d9567f\"},\"ciphertext\":\"40cce0a0b4eee4e4258341e18d880ff4b12ed25887252b90b7de4b6be86da5d5\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"4997f389a370424072967f6faeb709c238ff3cc3606a4101fc3dcbdd74b67108\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"97f0677a9f0ad0c6052896e1b5ae94f86025047b9fe2617c1fce220af0db0e82\"},\"id\":\"39626239-3338-4138-b662-363663333030\",\"version\":3}"
-  },
-  {
-    "privateKey": "629fd0230abafb7bc8d575c604e6a1b3fabb2c995d5d6d0ab504156ce22d7136",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0x8d991C6222F26B9aB960dbD28A1e9f29ED77272d\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"6177b7d2561056e9be3c991cce5148ce\"},\"ciphertext\":\"562c10f06d4fd050edcc4f99ef3390da34174c9d5a22ca46f4ed7d8ff0131e34\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"6c35d2b1918ff020bb52b4374105d9c5f9c20e586f2b4067818fe55ccc21372f\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"b19578d520eac69a989e196fa65b059be8dd5644e157d065a17379f85d2a2994\"},\"id\":\"66373132-6463-4634-b936-356330326230\",\"version\":3}"
-  },
-  {
-    "privateKey": "c6525c60b970082edf830ea2ebdbe2437bdf715e0bf7e3fbb8d4b0e73ec029bc",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0xF5D26831Be0B1fe5a6Ba5286DC59FaacE63f24aD\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"44d6acb359150e9019e1dda58da2ceca\"},\"ciphertext\":\"3e37afff8624b000591dd84745738c64c2728b94a74da60d8600fa9419381694\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"bbf3723986c6ce90f8fe7846d2bc0bef40885688bef01733158652019f7f6c22\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"6d8f5522a520dba0e6820715a4ae62dbc6412ae6d25aa3d2eacd8941f7c427a9\"},\"id\":\"35313363-3863-4265-b135-366531343061\",\"version\":3}"
-  },
-  {
-    "privateKey": "9a1975ac9e6669ea3e53248505312836cdfa187bdc356016aae13cc9937dfe7e",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0xB07Fc9BBBc3575f6002F6A071c025f09dC3b1600\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"98d1ab133dd8107b8d2ac751c6ad1739\"},\"ciphertext\":\"a06739df671a711b49e35963546a7364f26849c8999b1105e8445916285fb70a\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"5d5c19148d874db868aa707be7ea8018b16b53bb65a0f696b1b5c21b7be64dab\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"dacdf30263530d216b157c0d019b8ba7b2d28723c460bef453feae07f09d6507\"},\"id\":\"35636261-3930-4137-b363-376165663466\",\"version\":3}"
-  },
-  {
-    "privateKey": "83f3b281bd51c46909b097a516cf181ad049afe3d1e9830206fc784b3267c7cd",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0xc251Cd833942B0E3a315F1829b249FBb4270F01d\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"f45415e431c2b233db9741003fba78b3\"},\"ciphertext\":\"8a16a739928c60a5ae6fd8f27bbc32d93c6db298109432519a4c681243618e9a\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"5e504043c990517cceaa61858192682c44983bac500ea4d127c89a25cfde74fe\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"0d23e30d421c199304bfbdc2a14e2a717dfec64c7ed6494e9bbe276cf3f05082\"},\"id\":\"64313631-3530-4531-a235-633838376534\",\"version\":3}"
-  },
-  {
-    "privateKey": "b4825307a39a1a37d97faa6851337a5412f3b72541453acc774a195c98d1e164",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0xCE6C36Fff13EABD2936922D7a4e945491Da62744\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"8238351ae68d2b785f965fb070b489a8\"},\"ciphertext\":\"485eefcaa6f2e5c95e3fc504066be243cf104fca714ad37a0311705a71e11d5a\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"c25c274396608c47f369bcb492f4ffcf0c6130393b29bbbca1dd6c10a579b2f3\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"fbd963520c078000e769ee070ff1a6c3b88e139afb8b3933a0976240f1671f05\"},\"id\":\"36656433-6133-4338-b733-313166613965\",\"version\":3}"
-  },
-  {
-    "privateKey": "2e701a3e7294ecb097b711badc0bb0bd86d81c75a5ffb8fd41227d1028f59a1f",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0x9b32561E05aB55658f050C4cE53EDcd6A936D707\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"fc39dbd667e2b4bb643add88274438ff\"},\"ciphertext\":\"2c553f2f3f0d8032db0e2999601033e169775b5273cb3c22ae654e276a91ac43\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"93f93b57253f8c47bd5ea4ac59a0065f34c925581497bc5b136d5260564c8939\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"247acdebcf7faaa6179e17697d6a51564240e815af779e4cd243db299aca4fd5\"},\"id\":\"38373530-6430-4139-b631-653531333139\",\"version\":3}"
-  },
-  {
-    "privateKey": "3a08a854ab107b44005d8326526b361cf222517de5e9b3e8c045155956ad14e8",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0x25E8391dC1cf1ba5FD55341bec1B97a39542CF93\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"904c8c1daad999dfe9dce9bc4ebb7d4a\"},\"ciphertext\":\"c6c7d943de6362fd1bc9599021ea6cdcd8ad0c7c8e54b5c76645e17465e1d469\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"da258c156a05810328d1bad4de3232784c67db1bb4a6c07468051fbe5eb18ad1\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"89b0ce2fa23fa42eb41eebb4d7df55797221dfa2b4060ddbba1ed33447271987\"},\"id\":\"31643436-6236-4865-b462-636336366530\",\"version\":3}"
-  },
-  {
-    "privateKey": "b9a4685a64587cdf9a849b479ca6b12901dd363ecae487a20ab5aa69b09fc405",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0xa94658F5B0c902839657571fd879BCa0900C2eB9\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"638bbb0e6cc096e7ef5fd388916208ba\"},\"ciphertext\":\"f90c9ae5acf55578a71db71a34070685fadc9acb6b25e088b7a54d5747854eb7\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"07290fedf921b8003606158fcb479dcae2259311ce0da7f454335b3660f2ab2a\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"8de4dfa0a5b03077f40fb2b8e6ebd8e71c28d7a37190f60081c51307f922c1ae\"},\"id\":\"32363631-3663-4930-a236-306134616465\",\"version\":3}"
-  },
-  {
-    "privateKey": "28ff76ac522a1a1b4137d39060e15879b38ce1a2dde0ea92ac999da7e28c1cea",
-    "passphrase": "strong_password",
-    "keystore": "{\"address\":\"0x2F42cd4c0313f1A9B9AbE08Cfb18eA1808a367cF\",\"crypto\":{\"cipher\":\"aes-128-ctr\",\"cipherparams\":{\"iv\":\"ba302b5d9ea4435133801ea22a537e82\"},\"ciphertext\":\"ca21965ece6f3ddc8e533a80eea70e24258517250e528a4644dc511be6e90802\",\"kdf\":\"scrypt\",\"kdfparams\":{\"salt\":\"7f28e4b13034d8f69a3294eaa6f2d7187a3789600fd9fc2c1bcc2a2ddf1d24b8\",\"n\":8192,\"c\":262144,\"r\":8,\"p\":1,\"dklen\":32},\"mac\":\"a2b178f5d7dbd7201cc3199d9648eba794fded8c2156f7e3b6f61023d6a4be37\"},\"id\":\"38353362-3563-4539-b134-356661663937\",\"version\":3}"
-  }
-]
