--- packages/zilliqa-js-crypto/test/address.fixtures.ts
@@ -1,79 +1,0 @@
-//  Copyright (C) 2018 Zilliqa
-//
-//  This file is part of Zilliqa-Javascript-Library.
-//
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU General Public License for more details.
-//
-//  You should have received a copy of the GNU General Public License
-//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-export const addresses = [
-  {
-    public:
-      '0314738163B9BB67AD11AA464FE69A1147DF263E8970D7DCFD8F993DDD39E81BD9',
-    private: 'B4EB8E8B343E2CCE46DB4E7571EC1D9654693CCA200BC41CC20148355CA62ED9',
-    address: '4BAF5FADA8E5DB92C3D3242618C5B47133AE003C',
-  },
-  {
-    public:
-      '034CE268AC5A340038D8ACEBBDD7363611A5B1197916775E32481F5D6B104FAF65',
-    private: 'FD906F4A20A0507813FCB0D8D166CF45C1B17F217B776A003B3C3CC628C7E513',
-    address: '448261915A80CDE9BDE7C7A791685200D3A0BF4E',
-  },
-  {
-    public:
-      '02FA7A501F323CC53E070C0A945370368679E7572960EC24D8A0387EF3B50A2285',
-    private: 'C1FE69C394EE8564D0EA525FF74E5F37DF5267CF5C74F2E224757B7D9E594B73',
-    address: 'DED02FD979FC2E55C0243BD2F52DF022C40ADA1E',
-  },
-  {
-    public:
-      '036F8676E473AF20B4CCE7D327103DE4504A9C00EAE7EA03D0A365FB48817AC97F',
-    private: 'F197920FF83E3BD727560E1E08A4825910E3AA772B40EC4DEFFE241EE3A51D20',
-    address: '13F06E60297BEA6A3C402F6F64C416A6B31E586E',
-  },
-  {
-    public:
-      '034C39363529C2D4078F72B8C498C4CBC5BA5E10D8666FE06F104A27E0E44242A0',
-    private: '3D0C4009DA5B4630D292E93FC8DF7A55E9B8B31E182A50D113DE6E7A7A728B5B',
-    address: '1A90C25307C3CC71958A83FA213A2362D859CF33',
-  },
-  {
-    public:
-      '026DA5BF6C3A3C91C08A3DD7FBCE0BEAF5D436039C97B07A0F61AA4B9493E46787',
-    private: '8A06825735F2F374E5F60F0E08E06EF7169E2AE47532B741700574B9E1603DC0',
-    address: '625ABAEBD87DAE9AB128F3B3AE99688813D9C5DF',
-  },
-  {
-    public:
-      '032A661F9D4AB8DD9818CED2F62F3DA14FDD23E68E58D01A4AE186231D7FB609BB',
-    private: '715AB4286ECCB4E8E3AEC0AA8B6FB0E13C105B727E89AAE449F87F5DF4CFA53F',
-    address: '36BA34097F861191C48C839C9B1A8B5912F583CF',
-  },
-  {
-    public:
-      '036A94E925BB200DCE41A46C9026E023B226DD5B4CB227CE766D60CC8FAB218148',
-    private: 'B9056706EF197125598E7D5CA5AC9C47339B5C894C0656D33113052470BAF149',
-    address: 'D2453AE76C9A86AAE544FCA699DBDC5C576AEF3A',
-  },
-  {
-    public:
-      '0247F13639C7597C8AE2467DD29D219C2749260F690D4069930FDEB7866B2BC1FA',
-    private: 'A6DEDD1BE3A36945517A6C8A0E1353414DEBC04C50C9E036361C2A2D95EEF947',
-    address: '72220E84947C36118CDBC580454DFAA3B918CD97',
-  },
-  {
-    public:
-      '02883100F00EAB9FB6FD79034BBA84235C22ED642470FF6C1DB7CF7F782902D9DC',
-    private: '57489B87A3E726267EC5150FAB6E202694E806EC8FBBE386572CDE640BB3EE4B',
-    address: '50F92304C892D94A385CA6CE6CD6950CE9A36839',
-  },
-];
