+++ packages/zilliqa-js-crypto/test/address.spec.ts
@@ -15,9 +15,52 @@
 //  You should have received a copy of the GNU General Public License
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import { addresses } from './address.fixtures';
-import { checksummedStore } from './checksum.fixtures';
 import * as crypto from '../src/index';
+
+export const addresses = [
+  {
+    public:
+      '0314738163B9BB67AD11AA464FE69A1147DF263E8970D7DCFD8F993DDD39E81BD9',
+    private: 'B4EB8E8B343E2CCE46DB4E7571EC1D9654693CCA200BC41CC20148355CA62ED9',
+    address: '4BAF5FADA8E5DB92C3D3242618C5B47133AE003C',
+  },
+  {
+    public:
+      '034CE268AC5A340038D8ACEBBDD7363611A5B1197916775E32481F5D6B104FAF65',
+    private: 'FD906F4A20A0507813FCB0D8D166CF45C1B17F217B776A003B3C3CC628C7E513',
+    address: '448261915A80CDE9BDE7C7A791685200D3A0BF4E',
+  },
+  {
+    public:
+      '02FA7A501F323CC53E070C0A945370368679E7572960EC24D8A0387EF3B50A2285',
+    private: 'C1FE69C394EE8564D0EA525FF74E5F37DF5267CF5C74F2E224757B7D9E594B73',
+    address: 'DED02FD979FC2E55C0243BD2F52DF022C40ADA1E',
+  },
+];
+
+const checksummedStore = [
+  {
+    original: '4BAF5FADA8E5DB92C3D3242618C5B47133AE003C',
+    zil: '0x4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
+    zil_no0x: '4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
+    eth: '0x4bAf5fada8E5dB92C3D3242618C5b47133ae003c',
+    eth_no0x: '4bAf5fada8E5dB92C3D3242618C5b47133ae003c',
+  },
+  {
+    original: '4BAF5FADA8E5DB92C3D3242618C5B47133AE003C',
+    zil: '0x4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
+    zil_no0x: '4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
+    eth: '0x4BaF5fADa8E5Db92c3D3242618c5b47133Ae003c',
+    eth_no0x: '4BaF5fADa8E5Db92c3D3242618c5b47133Ae003c',
+  },
+  {
+    original: '448261915A80CDE9BDE7C7A791685200D3A0BF4E',
+    zil: '0x448261915a80cdE9BDE7C7a791685200D3A0bf4E',
+    zil_no0x: '448261915a80cdE9BDE7C7a791685200D3A0bf4E',
+    eth_no0x: '448261915a80CDe9bde7C7A791685200d3A0BF4e',
+    eth: '0x448261915a80CDe9bde7C7A791685200d3A0BF4e',
+  },
+];
 
 const bech32Tests = [
   {
