--- packages/zilliqa-js-crypto/test/checksum.fixtures.ts
@@ -1,96 +1,0 @@
-//  Copyright (C) 2018 Zilliqa
-//
-//  This file is part of Zilliqa-Javascript-Library.
-//
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU General Public License for more details.
-//
-//  You should have received a copy of the GNU General Public License
-//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-export const checksummedStore = [
-  {
-    original: '4BAF5FADA8E5DB92C3D3242618C5B47133AE003C',
-    zil: '0x4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
-    zil_no0x: '4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
-    eth: '0x4bAf5fada8E5dB92C3D3242618C5b47133ae003c',
-    eth_no0x: '4bAf5fada8E5dB92C3D3242618C5b47133ae003c',
-  },
-  {
-    original: '4BAF5FADA8E5DB92C3D3242618C5B47133AE003C',
-    zil: '0x4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
-    zil_no0x: '4BAF5faDA8e5Db92C3d3242618c5B47133AE003C',
-    eth: '0x4BaF5fADa8E5Db92c3D3242618c5b47133Ae003c',
-    eth_no0x: '4BaF5fADa8E5Db92c3D3242618c5b47133Ae003c',
-  },
-  {
-    original: '448261915A80CDE9BDE7C7A791685200D3A0BF4E',
-    zil: '0x448261915a80cdE9BDE7C7a791685200D3A0bf4E',
-    zil_no0x: '448261915a80cdE9BDE7C7a791685200D3A0bf4E',
-    eth_no0x: '448261915a80CDe9bde7C7A791685200d3A0BF4e',
-    eth: '0x448261915a80CDe9bde7C7A791685200d3A0BF4e',
-  },
-  {
-    original: 'DED02FD979FC2E55C0243BD2F52DF022C40ADA1E',
-    zil: '0xDed02fD979fC2e55c0243bd2F52df022c40ADa1E',
-    zil_no0x: 'Ded02fD979fC2e55c0243bd2F52df022c40ADa1E',
-    eth_no0x: 'DED02FD979fC2e55c0243Bd2f52DF022C40aDa1E',
-    eth: '0xDED02FD979fC2e55c0243Bd2f52DF022C40aDa1E',
-  },
-  {
-    original: '13F06E60297BEA6A3C402F6F64C416A6B31E586E',
-    zil: '0x13F06E60297bea6A3c402F6f64c416A6b31e586e',
-    zil_no0x: '13F06E60297bea6A3c402F6f64c416A6b31e586e',
-    eth_no0x: '13f06E60297bEA6A3C402F6F64c416a6B31e586e',
-    eth: '0x13f06E60297bEA6A3C402F6F64c416a6B31e586e',
-  },
-  {
-    original: '1A90C25307C3CC71958A83FA213A2362D859CF33',
-    zil: '0x1a90C25307C3Cc71958A83fa213A2362D859CF33',
-    zil_no0x: '1a90C25307C3Cc71958A83fa213A2362D859CF33',
-    eth_no0x: '1a90c25307c3Cc71958A83fa213a2362D859cF33',
-    eth: '0x1a90c25307c3Cc71958A83fa213a2362D859cF33',
-  },
-  {
-    original: '625ABAEBD87DAE9AB128F3B3AE99688813D9C5DF',
-    zil: '0x625ABAebd87daE9ab128f3B3AE99688813d9C5dF',
-    zil_no0x: '625ABAebd87daE9ab128f3B3AE99688813d9C5dF',
-    eth_no0x: '625aBAEBd87Dae9AB128F3b3ae99688813d9C5Df',
-    eth: '0x625aBAEBd87Dae9AB128F3b3ae99688813d9C5Df',
-  },
-  {
-    original: '36BA34097F861191C48C839C9B1A8B5912F583CF',
-    zil: '0x36Ba34097f861191C48C839c9b1a8B5912f583cF',
-    zil_no0x: '36Ba34097f861191C48C839c9b1a8B5912f583cF',
-    eth_no0x: '36BA34097f861191c48c839c9B1A8B5912f583cf',
-    eth: '0x36BA34097f861191c48c839c9B1A8B5912f583cf',
-  },
-  {
-    original: 'D2453AE76C9A86AAE544FCA699DBDC5C576AEF3A',
-    zil: '0xD2453Ae76C9A86AAe544fca699DbDC5c576aEf3A',
-    zil_no0x: 'D2453Ae76C9A86AAe544fca699DbDC5c576aEf3A',
-    eth_no0x: 'D2453AE76c9a86AAE544FCa699DBdC5C576aEf3A',
-    eth: '0xD2453AE76c9a86AAE544FCa699DBdC5C576aEf3A',
-  },
-  {
-    original: '72220E84947C36118CDBC580454DFAA3B918CD97',
-    zil: '0x72220e84947c36118cDbC580454DFaa3b918cD97',
-    zil_no0x: '72220e84947c36118cDbC580454DFaa3b918cD97',
-    eth_no0x: '72220E84947c36118CDbc580454DfaA3B918cd97',
-    eth: '0x72220E84947c36118CDbc580454DfaA3B918cd97',
-  },
-  {
-    original: '50F92304C892D94A385CA6CE6CD6950CE9A36839',
-    zil: '0x50f92304c892D94A385cA6cE6CD6950ce9A36839',
-    zil_no0x: '50f92304c892D94A385cA6cE6CD6950ce9A36839',
-    eth_no0x: '50f92304c892d94A385Ca6ce6cD6950ce9A36839',
-    eth: '0x50f92304c892d94A385Ca6ce6cD6950ce9A36839',
-  },
-];
