--- packages/zilliqa-js-crypto/test/keypairs.fixtures.ts
@@ -1,79 +1,0 @@
-//  Copyright (C) 2018 Zilliqa
-//
-//  This file is part of Zilliqa-Javascript-Library.
-//
-//  This program is free software: you can redistribute it and/or modify
-//  it under the terms of the GNU General Public License as published by
-//  the Free Software Foundation, either version 3 of the License, or
-//  (at your option) any later version.
-//
-//  This program is distributed in the hope that it will be useful,
-//  but WITHOUT ANY WARRANTY; without even the implied warranty of
-//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//  GNU General Public License for more details.
-//
-//  You should have received a copy of the GNU General Public License
-//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-export const pairs = [
-  {
-    private: 'b776d8f068d11b3c3f5b94db0fb30efea05b73ddb9af1bbd5da8182d94245f0b',
-    public:
-      '04cfa555bb63231d167f643f1a23ba66e6ca1458d416ddb9941e95b5fd28df0ac513075403c996efbbc15d187868857e31cf7be4d109b4f8cb3fd40499839f150a',
-    digest: '0e3e927f8be54eb20f4f47baa2f4d23649433591ea1b967127b971de2288609e',
-  },
-  {
-    private: '24180e6b0c3021aedb8f5a86f75276ee6fc7ff46e67e98e716728326102e91c9',
-    public:
-      '04163fa604c65aebeb7048c5548875c11418d6d106a20a0289d67b59807abdd299d4cf0efcf07e96e576732dae122b9a8ac142214a6bc133b77aa5b79ba46b3e20',
-    digest: 'f5b2dfddfdbd0d13a67085a5ad5744c0c6246d9704592116fccbf41978fe99c8',
-  },
-  {
-    private: 'af71626e38926401a6d2fd8fdf91c97f785b8fb2b867e7f8a884351e59ee9aa6',
-    public:
-      '0485c34ff11ea1e06f44d35afe3cc1748b6b122bb06df021a4767db4ef5fbcf1cdbed73c821d89724b59a0fc5e2f2e3e1a9c121d698fff36d750ee463117438a14',
-    digest: 'f8b9842a5f695139d22f49d1f706173b76fe0b5f680779558f23635a35425154',
-  },
-  {
-    private: 'b94289721618d1a8100bea8502fee149bae7313fcbaebf5ad6a867d557e82971',
-    public:
-      '04c7d47ad99dd1db85c9d0b6abe89ffc137f230c991a33890c05436a14974543a29d58bc4655269db6bcd6ea27adca287f956355c64d676cefcfea14051b5239c9',
-    digest: 'f6dae5bbe10084a0afcf363d11b08e4f6941d09dfbd39ce6474e115ff7908e05',
-  },
-  {
-    private: 'c577711506abf9dbdabd09c5ae66492e77d6450e8d739ce6493728e781150236',
-    public:
-      '0484929ab70d2b39703319a94144802ea9b9a7c8c1a673b10f01738db5b5c40ea8f534bc21ef1b167ffcac7d1a087ed5be2a5d1fa720bc016fbfbd9fb869aee4c6',
-    digest: '688c722f9c72cc4d760fd5967b3e26c3628bb4b3eb4b397ab7a2c3f26d312931',
-  },
-  {
-    private: 'e9521f5f9f2ac16f4e22681d43f2c1e15163ffe56e5912d57540dd59d71f2877',
-    public:
-      '04db0535d0e4337e5ada6d87b07569308f3f3de19aad15ba1f95108e96f855e17195e5b86773f7570645832048cb5e13e1721784ecce6bf69a0342f873e3afef58',
-    digest: '80a0001162a8230c4d3b624f8acf90b9c4d9d8f625703936a1e02aecb7d84d41',
-  },
-  {
-    private: 'def6ac38d746349abd8a28903c6b98b06156f6f3925d1c32dec9964541fbbfa1',
-    public:
-      '04dc5ca31dff767cb18f92d208fd74e36e800508d3b88bee61680bb7cefa0515943f2801d25bbce9c6a65721427f2443532980a0f7b7aae71b1fe0501179b8f641',
-    digest: 'c3f72427150375b7f9838f1a82873a66fcff822f529ad313e88c7785726f3ec1',
-  },
-  {
-    private: '93fe1b32f4f1358d6809982bce53b9e95c083eb889d14b0c2d0f96012074dc6b',
-    public:
-      '04987f1f6a2fe56cd0c2461f2920e3e0d9bd9de96a8febf9246e187e057c6141c494b697455dccefa29361b976698594082ed962da42bce09ef0ce59bd05510dd4',
-    digest: 'a26c10e3afcaa7aa20f21ba513747d1a9ba8c1b91b32bfabe88072e9519bf0f9',
-  },
-  {
-    private: 'efaa6048789fa0282302a5620a2fb0e2c60095a54c42b8312646bbbe1d2ec801',
-    public:
-      '04645a3995ff3309a36d1ec0991fd87ece2ee36e8887f0d5b304accb9152fe7a603aba24fefa904741e0fd67db24432a38153ecdc496ceec38eab1b7c2d8405de9',
-    digest: '5be58c9bc87b284af73274dc6dc01e38faaa25dfac9cbeb8941ea90c177d2cd8',
-  },
-  {
-    private: 'ca7de577b3e6968da27088d22e918c039a96af3d4821b7e103560fb6ca1185c4',
-    public:
-      '04f3e03a4ac451a78254fa3d056448db6cbc29d2ef429228c8435141a126c0c07f2446c68270016322e93a6a14f2e7c557bfdcc74e122a013498b7f1ebe008cc30',
-    digest: '0179e96ddeb567686e47f71794256ddbdbbfd5ca865b26d6751c66c84aa1ef4b',
-  },
-];
