+++ packages/zilliqa-js-crypto/test/schnorr.spec.ts
@@ -19,7 +19,6 @@ import BN from 'bn.js';
 import { randomBytes } from 'crypto';
 import elliptic from 'elliptic';
 
-import { pairs } from './keypairs.fixtures';
 const schnorrVectors = [
   {
     msg: 'A7F1D92A82C8D8FE434D98558CE2B347171198542F112D0558F56BD68807999248336241F30D23E55F30D1C8ED610C4B0235398184B814A29CB45A672ACAE548E9C5F1B0C4158AE59B4D39F6F7E8A105D3FEEDA5D5F3D9E45BFA6CC351E220AE0CE106986D61FF34A11E19FD3650E9B7818FC33A1E0FC02C44557AC8AB50C9B2DEB2F6B5E24C4FDD9F8867BDCE1FF261008E7897970E346207D75E47A158298E5BA2F56246869CC42E362A02731264E60687EF5309D108534F51F8658FB4F080B7CB19EE9AEBD718CC4FA27C8C37DFC1ADA5D133D13ABE03F021E9B1B78CCBD82F7FF2B38C6D48D01E481B2D4FAF7171805FD7F2D39EF4C4F19B9496E81DAB8193B3737E1B27D9C43957166441B93515E8F03C95D8E8CE1E1864FAAD68DDFC5932130109390B0F1FE5CA716805F8362E98DCCAADC86ADBED25801A9A9DCFA6264319DDAFE83A89C51F3C6D199D38DE10E660C37BE872C3F2B31660DE8BC95902B9103262CDB941F77376F5D3DBB7A3D5A387797FC4819A035ECA704CEDB37110EE7F206B0C8805AAEBF4963E7C4708CE8D4E092366E71792A8A3B2BBCDEE321B3E15380C541EF0930888969F7457AFE18588826A419D58311C1784B5484EECDB393F6A0ACA11B91DF0866B500B8DEE501FD7EB9BCE09A17D74124B4605ADFC0777BED9816D8D7E8488544A18D8045CB3283B0A752B881B5F500FADB59010E63D',
@@ -45,61 +44,38 @@ const schnorrVectors = [
     r: '4B73D4D919D7B4DEF330391899EA02023851CABE044E34E18EAE3E10588CECCD',
     s: 'D5DE85C4BDEA5910DC36AEF5660774D65291322C1E87FDA0D00C864E8C5FED29',
   },
+];
+
+const pairs = [
   {
-    msg: '3B0A947831FB58284F8F291A80A79099C05DFA394FFC1E827CEF0786875E5B1495CC8671F3A126D34AEF9782FE8D0EB2207F5BC585764709AEA129C37CC07D124F2BD9B9E618E8222E7B836B483ADE190A5D770EEA6B5EFBBB11E6CB218F64456B868FD07AC777B79E7D81AE6C3093A776792692096803909A4AFADEB670CBFD79E586CE044DD78386DC1984A9050B0772DB3FD18925511545FC382893AB681473467A7DF761A1516FC4ABC894328EEF5F9FD1893335AC90005637DFCD3EBDDF694C6DEA300A272C4EE6931F5064E5B5EDD9F728D8F4930D3EE1BD5B999FAE032CB16937DD027ACE29668EA5B30B67A35F65769AD3FE1C6FC3BE15B24CA4730577A88262C27B1236963EC9D2E0F8DFC52A36B03C34A9D8742045E9BA73B9ED8B7FC901F3483ADDD0473E487472D94578255CDC64F2FC76DFB10E29FF737221BABEADDD56A8720AA2BB990D05AB7138AE47AD3EDBEE4AAE26271B35AD1B8606BC7727B77142E2EED03CC782CCA974FB1ACCD49FF39DFE9A6E5FCAEBA41358C22441D67AE6ED35A0908D21A15CBA603CF70C7C5FA4BDC5B18A311593F16F978E4841EDC7C35929B1DD06F4C9EF5C4350CF0596D4D8714A4A85159EE7649B99AE67B87CBCA763E1A7EF20002E7DBBB1AAB01B4FD16ABDDC66CEE248C839525E716FA4026D6F08BEC3220F0B2B471B76CD071A24A7B873C47F5811AEAB39EBA62633',
-    pub: '0377757CCE0CF6AD0D63034EA802591865847DFF755387EB2FECC9290DD991CE96',
-    priv: '007D7F3258B9A585D789847AB0B96CBB66D35B865614671965D455E6142B8522',
-    k: '6A7196F9E7D9759F9202C6A517F62246DAECCC39C34842E59F1FFA158ABD440E',
-    r: '4D502A7A286EF5AA4244C313B420521B8C235EB03F9A4695AF8844FAE27F3B8E',
-    s: '730D6A135C0476C76E630E740D2415557784AF433B2BE0DEB3B383C1CEE6AD02',
+    private: 'b776d8f068d11b3c3f5b94db0fb30efea05b73ddb9af1bbd5da8182d94245f0b',
+    public:
+      '04cfa555bb63231d167f643f1a23ba66e6ca1458d416ddb9941e95b5fd28df0ac513075403c996efbbc15d187868857e31cf7be4d109b4f8cb3fd40499839f150a',
+    digest: '0e3e927f8be54eb20f4f47baa2f4d23649433591ea1b967127b971de2288609e',
   },
   {
-    msg: 'F01452F9B588E48ADD1F0FD2D08BA7443C4B3931AD4B267DA34FA8E08405C0DCA185D91768764EEF365DC2EDB529B6D14F4D6AB430D8458D9FF5138218F99C361F0067ABA79AA91776FBD49DED08920DD40AA7DD526605252F70850E85D61E161C7A6B2E081FAC05EB1FF8C1ACCA4E43D84EBD6C33FA8BB6279B0A22E791CB23722F50F4C86B9DE8B17425A894996D0BEE82B2BED453BDD118071832C0F1490D13E0F25A69F7D33E764C3F97DA57D99F3C8C296A2B89D6D20A6C710846B637FE022FDA54FC564DCE44AE4FFC423FC226FE2636D3DFA9CE1BE28C4E587E79CB100B36574C0F6275395973DFB7B0E35AB34F47E9D719F66E6B3C1DC7559E5B2628CDB255E150E78B3E96110B395DA5FFF2CF3E467E3E92D2B31DC373E9AC0F7044E31DB99D8BAC3F3DD7A01BB4AF17BD18F510849474B687799CFDF8DCEE7F21D220795D8CB51CCD43E91EB62FC6E39B6E473BC9CB54D829F16322F241240E34CF5D3A8C7E8EB9C9F0B0818A4B94150F05E80627F86831F4AEA68B856054D72568403693175F97D0C70409126A6DF69CB73DC15E3C3C47DF7F69ED1D3072FBDFEEC8D8961DA3E651352EC639BD5D43576FC937EDE1654EA5B8E3459D388CA599C8885F8218F8D5B882F86EA29626E2AD7B9FA391B90E8775B9E18020039263C6532490C332C92BB88CFFFDF010D4C79D39B7C85D21603937090FF0A9E3687CFDDF',
-    pub: '0269F42FDA412C290496073890D58B1B08A44C5C1C45C6A8846DF4A09A7C8EA7AE',
-    priv: 'B180861DD0ADD5FA0EAF0C2796F5991FBBC07AA3003CCD8656236EC1A4CD982A',
-    k: '98DB3EAB1BAA5568491AE7F9BCA17947C199D8BCD040B7C530B6E2B0EDA45392',
-    r: '28748E567B56680DA64792C70549BD4C512F8A3CA392442FC711DAEB3CE5E1FA',
-    s: '30E67D120C45BAECCF5EA38EC0DB4680457F0B56EADDABF64E00EB2B46C8DEAC',
+    private: '24180e6b0c3021aedb8f5a86f75276ee6fc7ff46e67e98e716728326102e91c9',
+    public:
+      '04163fa604c65aebeb7048c5548875c11418d6d106a20a0289d67b59807abdd299d4cf0efcf07e96e576732dae122b9a8ac142214a6bc133b77aa5b79ba46b3e20',
+    digest: 'f5b2dfddfdbd0d13a67085a5ad5744c0c6246d9704592116fccbf41978fe99c8',
   },
   {
-    msg: 'B53659B821AAF67FE7EDE7063BC1D761B19DC51059D4E309F719703234E3B86A2C48CD99CCA48CAC5C111AFB8FD7CBDD8A5E5B88476D07513DDBFF5EF490F079C7BB2002F98E6471A81D6DF17FC736F7A7C361A945B7FA8205BF5DCBE9304B2ED0D0F012C7036CA58C198E62F3B075C059ACBFF322E0526E6A1E2D24F344ECC68DC411B0F101C058E8C96521964B337625FF73C7C261B85401C9031A125768CB87558D28889FA4EC3D0C77334745B64187E2459E919B083EC846994F5CBC62D0FF160690B7B1B441CDC2E4A75B59AF5CF4A1308DA3A48F6FA70339BBBC16362A46CEDDC42AB7ADF185AD88A6FC94AD0A9A9E6F5605A8B0E03896957FC5987E29CA2BA17207D9F65803EA7E25AF9ABF93E0C3B1FB904D16B30C9082588AEDEF2338893D7E03C877D1DB50C4E1289D558C8386E1FEC452B395486B268BF6848F757AD6737F088B38079D79D280DC9EFD2621BD55B314B4B2D8B5FE5366ECFDB7FC1A5AFDB4B7C758FC25466813912B77318E1148CC45F7EDE7F393D4EFE2976F54270C9D6865B9AB65EA85F8FDE3620E1D1F88CFABB36C9208374A16433EE32FD5FA078822096A4629397BE42EB3CDA84C8A6432B913D814939D436E41419FDFA90E6352FAF7DF20951C204388FC3896D3B4F8B0ABD3AC90FD9DA4B2A8B917963C8A5CE8FD0704F7666742D04E3DF8D55FA2B690DB584C08C7F5005E50E91587B7',
-    pub: '031451E51EE540E0C27696597C6C26035E553ED2604B45ADA1965C754A4978EA22',
-    priv: '08842B1E517039D735B731893502D727FEA2AD0845EFB89E022B190ABDD07B87',
-    k: '7FB8B8E25FAA6CFF4256CA92C769FD370324D366D81C18E41C1E02CC0A1CCF57',
-    r: '11396DF5E72C53D3253D37290D03585A95765A011254E9E5445D8A55C036BA27',
-    s: 'DEC36C2E6C0AAA893FF732ACFBD7AB7811563F3D3091CBB5AFAF13104694D531',
+    private: 'af71626e38926401a6d2fd8fdf91c97f785b8fb2b867e7f8a884351e59ee9aa6',
+    public:
+      '0485c34ff11ea1e06f44d35afe3cc1748b6b122bb06df021a4767db4ef5fbcf1cdbed73c821d89724b59a0fc5e2f2e3e1a9c121d698fff36d750ee463117438a14',
+    digest: 'f8b9842a5f695139d22f49d1f706173b76fe0b5f680779558f23635a35425154',
   },
   {
-    msg: '903EAFB5D25D35FD684DEC82BBFAF5B00B2B00138910915FB01326717AE36F099EC94AC1F86667ADE7E1D11487999ED6321D990600B852286C3581CBF133A3047B21D2418A82667AB32F260C555B80E13B4254E4C75692823A3A06872CEF57E381D1A4DAE918330854A88B6B9D93FEC884BB64BF07C997F56FC66B7F02FA5DBE53CBBCB04901A49B85F97F1A2AF32F29D8DA8F8078E5D2DD5F814E0B06E7903DA7B27F52C6736B19DCCD446D4DCCF76986C87002B27372B64502A489304C3F2D4A6B3F7C568B9E1F1B7B983F8863D48DA3AA1B743F06922754EE8C759A06EA9D8B62F3E2F3507035EE203801A1DF3B6C2C2C9E302C9E374F0F371A9C267B6452E246A139E6703AC7C2429151330F980478145CD2DB37F707241AD4499892D1EFEB8718DEF25918BACDE7778ED02129A3B9EFB346B48C31E925AF212C7DF2FAC27C7042A43436621AAFCE86C34838FAD1E39B5B0B04B9DE8ED42EC90AE62628C2F9F173E8D40C9116249AD03B05D6BFA9761493D4A3791668E166526CF810E34FAB41C147ACF45679C85B2EE22C4B9CAE63ECE97F9ACD75EE71738FF836FF08089D7448405FD96EB1232274E8507AB8FB0202CAB2613B13B363DBF1D767EB053F651CD952D44DD23D283A73276B96A3C68A42DB61A9DBAF13EE613AF13BB77BDC57DACC8CE210A1DF110194F234A580E7564A19AB5586B0642497F18215AC639F',
-    pub: '0359864E21B3F2AE4FD6FA5F488721D46B5A2A3B55CAA2047826D4BA459A141919',
-    priv: '9CB708971279F51BD67F33D5CBE1E29D147A641F148ED7AFE761E644A71AACDE',
-    k: '227A67404D79C9B86C3047473107097F57F5984ADFB6AB0FEB8CEE21AB0A2B91',
-    r: '06B7422F90F6C7A2378F731007C3FFBD73B83B42126B072A00EF679F86138F35',
-    s: 'A7D5A1CC9BF22022445EED23128006BD6D72E0E391A6CEE26B0ACA0E8B65D1D2',
+    private: 'b94289721618d1a8100bea8502fee149bae7313fcbaebf5ad6a867d557e82971',
+    public:
+      '04c7d47ad99dd1db85c9d0b6abe89ffc137f230c991a33890c05436a14974543a29d58bc4655269db6bcd6ea27adca287f956355c64d676cefcfea14051b5239c9',
+    digest: 'f6dae5bbe10084a0afcf363d11b08e4f6941d09dfbd39ce6474e115ff7908e05',
   },
   {
-    msg: '286D4DEE8E9E1F7B8DCEC705E9CFCFFFB43C5F3D18F06F80A06A36396D56B7EFE81B00B047A88BAD1FDB1BBF307149E2953DA47B939370999CC440F218E1BD04CC2691C0C335332919BB5C515BED14B6B1FDE2417887FB013E690DCD5EF0C636E6A384F828C0393CFB699718C7B3E38FE59DC890D32C9FCC81494B6D56291582AA63E29926D8E3176B16EAE0411599BE3B188E6E6C06D8A4E96E6DD8422FD6FB794E7891309EA9F8D48AA7962F3512F57015BAC86EA96D292F6FD92E1EB513F00DB469117DFF8F1FED1528A2945A8CBFE953B7CF4E3B3BE98BAE5351F37F29F39ABF01457153D6360491F791449AECAB1C05DB6E60A2E8B2FD5C6D68F3364B855F5F42A82599591A2E0D9ACD6AB9F0CC4EBE9E0DF1ABB33320F8BBDDCA558BEA493F7CEACD7D04204F2004D55059E07C9D8BBA4EE4B4A8C46133F968DA63E4E409D05E825A4AA33744F419C37FA5BD0F26BE28604AC6ED23C35C5F3FBF33BBA2FE15F3284A68B1B1BD65B4478915CF3872BB539CA5B5EE84648CEA4B30C2D86C6CE4B9A530FD2582994790C3F8B8D73DE63AD3784394C3D4F99A4CC9C6F2D900C73E15908AD0FABFA8DCA09B9ED71D0010D4C4CEFB4EA7278785F1D55670D9673CACBD0BB17BFA4836D3A4ADB85033635F5B0D3CC0358435F3600D27B488614BFCDEA9D7B13474061075E2896E8B254AEBF8E7B076CF47F47027247DF6189301',
-    pub: '020F2842B5DD31CCD0EFBC41986F83EEE18E301EC87A679E59D7D99EA3260B5061',
-    priv: 'E6E7FDD7067D631CA8F5A20EEA4CBEE8474CA0A025B4F1CA50B2AD7E0E263680',
-    k: '4F16E1CB6AC95981DEB3584CBE1D1CD5069266AAAEA362BA47A4F8D4BD2F75DA',
-    r: '9F60EB3DE84BA67834690B32A8219DA3B56B1C83788D4451562B8956F373EA85',
-    s: '8BA219A50605778FB1C11DD45288DB362A573A4F5354ACF997D7A5AE86B00B3F',
-  },
-  {
-    msg: '3B7DC5890F6FB89C7B3E88B5EF8B742B209DF6A99900535AD32B1F980F7F5F2E13C6488B9D1CB8A0B198376E1C16505D9CB90B61CCCED35256E505417F550A7EA043CD7F55E94B5E5C18BBA6BE3F229727CAD0226F81355F23AD0124C8E9B1B7A3C86943A06B9796F380FF3154F787ACEA4193A8205400E6B7F0AC002BFDC410060D3AEF8AF278041695224351DD866014EF75AAD869E271329166597D9095E779D8CF18D75D0385636C58D74B94CFD705F650A6FE8C1B03179E5E950CAC438B82FE68CE8F2C57D3D851E7A4E3DB521D6BED195FCCB434D5A220857F82C750A5FAA758A6CF4055E7356635B58E9E5E565925424EDFAA12815F53C3ED33D28508FB763591C0A48309148667057E4FBA646454C3FB38EDA2F1FF038B3549ABC5BDA78FCC7DA2BD399FB7E4F132538933E482BCFFA4B5EE8ACE0CAF69A4B066C30ED1374C7BBD138425A03CB6C88125B02C8D9F14945F6C8CC2AA89A8AD8014D436D6805177DC81B6C381B2231069CDD23B0EA742CFB5A6DC89AE28A9F3882F0F5440A7C94DF2CD197DC9C14FC839E70708387A2364D248E74BE921571676A6C928D06739C022727AB8BA6C0EB9817FFCEBD4319B3D3BCBB3F3435BADFD367A234F471C9EDFC11E3B43D91F48E736AE17F630E0B9F6A7294388C25ACCD03A246A48226FC2A0B438D25F27CE272F21A8E4FEE9884142107C2D8C228FCDF70198AADA',
-    pub: '02A96092EB1B7FCBAA29D763690787330E181398910291716CCEF2D68DE0EC1D39',
-    priv: 'C9C3874EB57B232095C627546B23076E4F2F98C8CD79DDD1CA5A4B4E72CE86DC',
-    k: '5B062A88B9FC44447E3D6A3D32C58D523EF9CAE2D786F23D69F6D9DA2943D305',
-    r: '69883DE347B3B9D0FAC6CBEB94F9F569C22152E41B9A21BDFD6177999D93CF7D',
-    s: '39A0756F7C145844276183446F4CBB923BCE83F7EAD3C1D669270FF89EAA352E',
-  },
-  {
-    msg: '66878CAB8808B926767A5FF0AAD6F9DD0921FBAB01B7D52DB72E3F982AA7B33E01F1BB87D9767BCEFE311BAA27C9FD3495427E630EA4759BF6E1E3E1B83B70128916FD4753BB7CEBE5044D639D5AE30D4A8A18655B9A16F058BEC7391845C8B4E9166A4528821F663F00800C7A5E1614B3ABA4CC0B0C42A0CC955C6D8586ADBA5A7A819F2B1B9EACEE86E39CC6391187B242A6C624AAB0CD7A9DE63184FEEB181D54A1D26AF9D940B4CF3CADCF000FFCDB2916CD09B9B4C1B6E878555B3001913D5ADFD98104900E70537E63796B17A50E644BF624FCA3CC5BAB2B41686515DDEF9EA584C67F42B873A0EF6DB8E014040373C90F91942EF4DA5BA531A9F64039276A215993FC881F37DD17FDD78F3DF4137F691EDD3670917363BB5CEE1EDAB32FC2238D2ACCFECC729C7B53DAED40068ED539C73957C16C962CEB3FE8AE4DBD63CBFA6C799F618AA3691AE7B77C6F5D0A56D0031E019C1503D13EF899E91F5C680A9C3692C4463843443758978B3A780B94196CB03828DD62B4E545CE3D4EA1D5DB71CB92E3C7D35A56C5BDA5B718DCD45585210AB75C58D7FA54C22F50CACCE5905A8BB9BB2B49E0EB15853C8A742BC73505CAE5049712BB91B1ADA2D4F80F733EF705485069A117E09983B7A0B7F26F6E1004154A088F5E8B071BD31A3D0D372026394922B1411D9668991A5A27315BB0FC84E10542E9F8F0069B0E5FDCE6',
-    pub: '03EB2C93532FC170AEE356B7FC2FCF1C41C9C5CB59F0063C8DFC4C03A3C4AC1345',
-    priv: '0B663B0F9BB529F7D5FB12E7D6AD6C2ED7B2516692C174E482ACE7717DD6E714',
-    k: 'D4900AA69CBFDBD481C3D133FC3D1AF7F033BF3C0E5D3AE44C7B3854F2DAB37A',
-    r: '3594889AAF6E06513D7BD1087D263B398BA28B7B32A3520B8D9D33BAAA1876CD',
-    s: '5384923B3CB692DF7B72BDF12BDF888AD50EC623C84779B7DE40AB74B32F1332',
+    private: 'c577711506abf9dbdabd09c5ae66492e77d6450e8d739ce6493728e781150236',
+    public:
+      '0484929ab70d2b39703319a94144802ea9b9a7c8c1a673b10f01738db5b5c40ea8f534bc21ef1b167ffcac7d1a087ed5be2a5d1fa720bc016fbfbd9fb869aee4c6',
+    digest: '688c722f9c72cc4d760fd5967b3e26c3628bb4b3eb4b397ab7a2c3f26d312931',
   },
 ];
 
@@ -110,7 +86,7 @@ const secp256k1 = elliptic.ec('secp256k1');
 
 describe('schnorr', () => {
   it('should fail for bad signatures', () => {
-    schnorrVectors.forEach(({ priv, k }, idx) => {
+    schnorrVectors.forEach(({ priv, k }) => {
       const pub = secp256k1.keyFromPrivate(priv, 'hex').getPublic(true, 'hex');
       const badPrivateKey = pairs[0].private;
